* diagnostic.c (bt_callback): Cast pc when calling fprintf.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@191766 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/diagnostic.c
@@ -364,7 +364,7 @@ bt_callback (void *data, uintptr_t pc, const char *filename, int lineno,
     }
 
   fprintf (stderr, "0x%lx %s\n\t%s:%d\n",
-	   pc,
+	   (unsigned long) pc,
 	   function == NULL ? "???" : function,
 	   filename == NULL ? "???" : filename,
 	   lineno);
